actualizado avg a double

+++ TriaCultura_service/Models/TriaCulturaRepository.cs
@@ -105,13 +105,13 @@ namespace TriaCultura_service.Models
             return a;
         }
         
-        public static int getCountRating(int project_id)
+        public static double getCountRating(int project_id)
         {
             List<rating> ratings = context.ratings.Where(x => x.project_id == project_id).ToList();
             
             if (ratings.Count > 0)
             {
-                int total = (int)ratings.Select(x => x.rate).Sum();
+                double total = (double)ratings.Select(x => x.rate).Sum();
                 return total / ratings.Count;
             }
             return 0;
